Network error recovery, gui changes
The patcher should now keep retrying the connection if it fails due to
an IO error. (badly tested, but it is no worse than the previous version
even if my code fails)

The progress bar is now a lot nicer. Furthermore, the time remaining
indicator now only appears when a non-retarded estimate exists.